refactor: modularized
