+++ src/components/achievement.tsx
@@ -1,25 +1,46 @@
-export default function Achievemnet() {
+import { nameAcronoym } from '@/lib/utils';
+import { ImageType } from '@/types/global';
+
+import { Icon } from './icon';
+import { Avatar, AvatarFallback, AvatarGroup, AvatarImage } from './ui/avatar';
+
+export function Achievement({
+  title,
+  users,
+  ratingText,
+  star,
+}: {
+  title: string;
+  users: ImageType[];
+  ratingText: string;
+  star: number;
+}) {
   return (
-    <div className="w-full lg:w-auto flex-1 flex lg:items-center justify-center px-8 py-5 lg:py-16">
-      <div className="max-w-md w-full">
-        <div className="text-white mb-12">
-          <div className="text-sm mb-8 flex justify-center lg:justify-start items-center gap-2">
-            <span className="w-2 h-2 bg-[#ff6b2c] rounded-full"></span>
-            <span className="text-xs text-center lg:text-left lg:text-base opacity-70">
-              2,847 trusted customers working with this tool
-            </span>
-          </div>
-          <h1 className="text-center lg:text-left text-lg lg:text-5xl leading-tight mb-6">
-            The fastest tradie wins the job.
-            <br />
-            Are you fast enough?
-          </h1>
-          <p className="text-center lg:text-left text-sm lg:text-base opacity-80">
-            Stop chasing quotes and start closing jobs. Every minute counts when
-            making <span className="text-[#ff6b2c]">$3,000-$5,000</span> in one
-            visit alone.
-          </p>
+    <div className="flex flex-col justify-center items-center gap-2.5 w-full h-fit">
+      <p className="text-neutral-600 font-semibold">{title}</p>
+      <div className="flex justify-center gap-4 items-center w-full">
+        <AvatarGroup className="ring-white">
+          {users.map((avatar, i) => (
+            <Avatar className="ring-white ring-0.5 size-4" key={i}>
+              <AvatarImage src={avatar.src} alt={avatar.alt} />
+              <AvatarFallback>{nameAcronoym(avatar.alt)}</AvatarFallback>
+            </Avatar>
+          ))}
+        </AvatarGroup>
+        <div className="flex">
+          {Array.from({ length: star }).map((_, i) => (
+            <Icon
+              key={i}
+              name="Star"
+              fill="#FFD700"
+              stroke="#FFD700"
+              className="size-4"
+            />
+          ))}
         </div>
+        <span className="text-[0.75rem] text-neutral-600 font-semibold">
+          {ratingText}
+        </span>
       </div>
     </div>
   );
